♻️ extract logging from test server
